Restructured files to enable page routing
